Add basic webpack for JS bundle
This is only a first stab at adding webpack bundling for our JS here. I know we could load up more through this, like our styles, but I'm unsure of the immediate benefit to that.

Also want to use this as a brief opportunity to figure out if webpack is the right choice here, or if Rollup or something else is. I'm still familiarizing myself with some of the tooling and concerns around requiring the transpilers since the regular imports we have here today seem to work.